// AIRMAM5/Scripts/Models/Enum/Controller.ts
/**控制器名稱 */
export enum Controller {
    Account = 'Account',
    ArcPre = 'ArcPre',
    ArchiveMove = 'ArchiveMove',
    Ann = 'Ann',
    Booking = 'Booking',
    BatchBooking = 'BatchBooking',
    ColTemplate = 'ColTemplate',
    DocViewer = 'DocViewer',
    Delete = 'Delete',
    Dir = 'Dir',
    Error = 'Error',
    Group = 'Group',
    Home = 'Home',
    L_Log = 'L_Log',
    L_Login = 'L_Login',
    L_Search = 'L_Search',
    L_Tran = 'L_Tran',
    L_Upload = 'L_Upload',
    License = 'License',
    Materia = 'Materia',
    MyBooking = 'MyBooking',
    Notify = 'Notify',
    Remind = 'Remind',
    Report = 'Report',
    Rule = 'Rule',
    Shared = 'Shared',
    Subject = 'Subject',
    SubjExtend='SubjExtend',
    SysCode = 'SysCode',
    Synonym = 'Synonym',
    Search = 'Search',
    Tsm = 'Tsm',
    UserCode = 'UserCode',
    User = 'User',
    VerifyBooking = 'VerifyBooking',
}