// AIRMAM5/Scripts/Models/Enum/TabNameEnum.ts
/**
 * 分頁名稱列舉
 */
export const enum TabNameEnum {
    /**段落描述 */
    Paragraph = 'paraDescription',
    /**媒體資料 */
    MetaData = 'metadata',
    /**關鍵影格 */
    KeyFrame = 'keyframe',
    /**基本資訊 */
    BaseMeta = 'baseMeta',
    /**詳細資訊 */
    DetailMeta = 'detailMeta',
    /**聲音專輯資訊 */
    albumInfo = 'albumInfo',
    /**圖片資訊 */
    exifInfo = 'exifInfo',
    /**文件資訊 */
    documentInfo = 'documentInfo',
}
/**
 * 分頁名稱中文列舉
 */
export enum TabNameChineseEnum {
    /**段落描述 */
    段落描述 = 'paraDescription',
    /**媒體資料 */
    媒體資料 = 'metadata',
    /**關鍵影格 */
    關鍵影格 = 'keyframe',
    /**基本資訊 */
    基本資訊 = 'baseMeta',
    /**詳細資訊 */
    詳細資訊 = 'detailMeta',
    /**聲音專輯資訊 */
    聲音專輯資訊 = 'albumInfo',
    /**圖片資訊 */
    圖片資訊 = 'exifInfo',
    /**文件資訊 */
    文件資訊 = 'documentInfo',
}